+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v0.9.0
+
+- Introduced Conversation Handler
+- [breaking change] added error return type to handler functions
+
 ## v0.8.0
 
 - Changed poll scheduling setup to allow for more frequent polls